move average rating into callbacks and components
