+++ src/callbacks.py
@@ -3,6 +3,7 @@ import pandas as pd
 import ast
 import altair as alt
 import json
+import plotly.graph_objects as go
 
 # ingredient icons callbacks
 csv_path = "data/processed/processed_cookie_data.csv"
@@ -169,3 +170,72 @@ def update_recipe_count(rating_range=[0, 1], selected_ingredients=None):
     num_recipes = filtered_df["Recipe_Index"].nunique()
 
     return f"{num_recipes}"
+
+# average rating callback
+df = pd.read_csv("data/processed/processed_cookie_data.csv")
+
+@callback(
+    Output("rating_gauge", "figure"),
+    Input("rating_gauge", "id"),  # No need for slider input
+    Input("rating-range", "value"),
+    Input("ingredient-checklist", "value"),
+)
+def update_gauge_chart(_, rating_range=[0, 1], selected_ingredients=None):
+    """
+    Compute the average rating and update the gauge with a moving dial color.
+    """
+    # connect to the ratings slider
+    filtered_df = df.query('Rating.between(@rating_range[0], @rating_range[1])')
+
+    # If ingredients are selected, filter by those as well
+    if selected_ingredients:
+        filtered_df = filtered_df[filtered_df["Ingredient"].isin(selected_ingredients)]
+
+    # group by recipe ID so that there is only one entry per recipe instead of per ingredient in the recipe
+    # the recipe's rating per ingredient should be the same, so "mean" doesn't really do anything
+    filtered_df = filtered_df.groupby("Recipe_Index")['Rating'].mean().reset_index()
+
+    # Compute average rating
+    avg_rating = filtered_df["Rating"].mean() if not filtered_df.empty else 0
+
+    # Create Plotly Gauge with a dynamically moving dial color
+    fig = go.Figure(
+        go.Indicator(
+            mode="gauge+number",
+            value=round(avg_rating, 2),
+            # title={"text": f"Average Rating: {avg_rating:.2f}", "font": {"size": 16}},
+            domain={"x": [0, 1], "y": [0, 1]},   # Fill the entire chart area (full circle)
+            gauge={
+                "axis": {
+                    "range": [0, 1],
+                    "tickmode": "linear",
+                    "tick0": 0,
+                    "dtick": 0.2,
+                    "tickfont": {"color": "#000", "size": 12}
+                },
+                "bar": {"color": "#3E2723", "thickness": 0.3},  # Ensure the dial color is distinct
+                "steps": [
+                    {"range": [0, avg_rating], "color": "#906A51"},  # Color up to average rating
+                    {"range": [avg_rating, 1], "color": "#F5E1C8"},  # Remaining range
+                ],
+                "threshold": {
+                    "line": {"color": "#3E2723", "width": 4},
+                    "thickness": 0.75,  # Adjusted for a clear marker
+                    "value": avg_rating
+                },
+                "borderwidth": 0,
+                "bordercolor": "#D2A679",
+                
+            }
+        )
+    )
+
+    # Adjust layout to fit the small container
+    fig.update_layout(
+        autosize=True, 
+        margin=dict(l=5, r=5, t=35, b=5),
+        paper_bgcolor="#D2A679",  # Match outer container
+        font=dict(color="#000")
+    )
+
+    return fig
